Adapter Factura Detalle

+++ app/src/main/java/uca/ni/edu/kelani/bd/entidades/FacturaDet.kt
@@ -5,5 +5,5 @@ data class FacturaDet(
     val id_factura_det:Int,
     val id_producto:Int,
     val cantidad:Int,
-    val subtotal:Int
+    val subtotal:Double
 )
